working engine initializer
